rename FlexibleDecay_settings file
