+++ src/module.mk
@@ -47,7 +47,7 @@ LIBFLEXI_SRC := \
 		$(DIR)/slha_io.cpp \
 		$(DIR)/spectrum_generator_problems.cpp \
 		$(DIR)/spectrum_generator_settings.cpp \
-		$(DIR)/FlexibleDecay_settings.cpp \
+		$(DIR)/decays/flexibledecay_settings.cpp \
 		$(DIR)/string_conversion.cpp \
 		$(DIR)/string_format.cpp \
 		$(DIR)/string_utils.cpp \
@@ -138,7 +138,7 @@ LIBFLEXI_HDR := \
 		$(DIR)/slha_io.hpp \
 		$(DIR)/spectrum_generator_problems.hpp \
 		$(DIR)/spectrum_generator_settings.hpp \
-		$(DIR)/FlexibleDecay_settings.hpp \
+		$(DIR)/decays/flexibledecay_settings.hpp \
 		$(DIR)/string_conversion.hpp \
 		$(DIR)/string_format.hpp \
 		$(DIR)/string_utils.hpp \
