+++ templates/decays/decays.hpp.in
@@ -39,7 +39,7 @@
 #include "wrappers.hpp"
 #include "error.hpp"
 #include "physical_input.hpp"
-#include "FlexibleDecay_settings.hpp"
+#include "decays/flexibledecay_settings.hpp"
 
 namespace flexiblesusy {
