+++ templates/librarylink.cpp.in
@@ -36,7 +36,7 @@
 #include "problems_format_mathlink.hpp"
 #include "slha_io.hpp"
 #include "spectrum_generator_settings.hpp"
-#include "FlexibleDecay_settings.hpp"
+#include "decays/flexibledecay_settings.hpp"
 #include "standard_model_two_scale_model.hpp"
 #include "lowe.h"
 @decaysIncludes@
