+++ templates/run_cmd_line.cpp.in
@@ -23,7 +23,7 @@
 #include "@ModelName@_observables.hpp"
 #include "@ModelName@_slha_io.hpp"
 #include "@ModelName@_spectrum_generator.hpp"
-#include "FlexibleDecay_settings.hpp"
+#include "decays/flexibledecay_settings.hpp"
 @decaysIncludes@
 
 @solverIncludes@
